requirements.txt
rembg
opencv-python
tqdm
pillow
numpy
onnxruntime
pytest
